// src/app/api/model/ticket-status.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum TicketStatus {
    OPEN = 'OPEN',
    IN_PROGRESS = 'IN_PROGRESS',
    WAITING_FOR_CUSTOMER = 'WAITING_FOR_CUSTOMER',
    RESOLVED = 'RESOLVED',
    CLOSED = 'CLOSED'
}